Rework attribute system again.
Still not sure that I have it right.  Moved to treating things as just bags of
data instead of objects.  Makes it easier to manipulate them.

Still open questions about circular references in dynamic attributes and how
that all works.  What are the inputs to those? What is the evaluation context? I
think eventually we'll end up with something more sophisticated.

Signed-off-by: Joe Beda <joe.github@bedafamily.com>